+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
-- Update descriptions and readme.
+- Update descriptions and README.
+- Add example.
 
 ## 1.0.0