+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+- Add indexable to Account.
+
 ## 1.0.1
 
 - Update descriptions and README.